pyproject.toml
[tool.poetry]
name = "app"
version = "0.1"
description = "fastAPI boilerplate"
authors = ["Raheel Siddiqui"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.85.0"
FastAPI-SQLAlchemy = "^0.2.1"
orjson = "^3.8.0"
pydantic = "^1.10.2"
python-dotenv = "^0.21.0"
requests = "^2.28.1"
SQLAlchemy = "^1.4.41"
uvicorn = "^0.18.3"
black = "^22.8.0"
alembic = "^1.8.1"
pytest = "^7.1.3"
unittest2 = "^1.1.0"
psycopg2 = "^2.9.3"
loguru = "^0.6.0"
tenacity = "^8.0.1"